task/requirements.txt
pep8==1.7.1
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
pytest>=7.1.3
numpy>=1.23.4
Pillow>=9.2.0
async-asgi-testclient